+++ CHANGELOG.md
@@ -13,6 +13,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.0.0-rc.197 (2021-??-??)
 
+#### :rocket: New Feature
+
+* Now all watchers support suspending
+
 #### :memo: Documentation
 
 * Added documentation `iDynamicPage`